python-weather-station/requirements.txt
msgpack-rpc-python
certifi
paho-mqtt
./kpn_senml-1.1.1-py3-none-any.whl
